Dont remove input files on failure

+++ dias_apis/coah/coah.py
@@ -7,6 +7,7 @@ EO data access - COAH API
 
 import os
 import time
+import shutil
 import requests
 from requests.status_codes import codes
 from zipfile import ZipFile
@@ -142,10 +143,14 @@ def do_download(auth, product, env, max_attempts=4, wait_time=30):
         except Exception as e:
             log(env["General"]["log"], "Failed download attempt {} of {}: {}".format(attempt + 1, max_attempts, e), indent=1)
             try:
-                Path(file_temp).unlink()
+                if os.path.exists(product_path):
+                    shutil.rmtree(product_path)
+                if os.path.exists(file_temp):
+                    Path(file_temp).unlink()
             except:
                 pass
             time.sleep(wait_time)
+    raise ValueError("Failed to download file after {} attempts".format(max_attempts))
 
 
 def authenticate(env):
